Bugs bugs ...

+++ controllers/crud.go
@@ -1,13 +1,13 @@
 /**
- * Initialize base CRUD controller class.  Malina eCommerce application
+ * Initialize base CRUD controller class.  github.com/ilyaran/Malina eCommerce application
  *
  *
- * @author		John Aran (Ilyas Toxanbayev)
+ * @author		John Aran (Ilyas Aranzhanovich Toxanbayev)
  * @version		1.0.0
  * @based on
  * @email      		il.aranov@gmail.com
  * @link
- * @github      	https://github.com/ilyaran/Malina
+ * @github      	https://github.com/ilyaran/github.com/ilyaran/Malina
  * @license		MIT License Copyright (c) 2017 John Aran (Ilyas Toxanbayev)
  */
 
@@ -16,13 +16,14 @@ package controller
 import (
 	"net/http"
 	"regexp"
-	"Malina/config"
+	"github.com/ilyaran/Malina/config"
 	"strings"
-	"Malina/libraries"
+	"github.com/ilyaran/Malina/libraries"
 	"strconv"
 	"github.com/gorilla/mux"
 	"fmt"
-	"Malina/models"
+	"github.com/ilyaran/Malina/models"
+	"github.com/ilyaran/Malina/views"
 	"html/template"
 )
 
@@ -50,41 +51,42 @@ import (
 
 **/
 
-
 type CrudController struct {
-	dbtable  string
-	order_by string
+	dbtable       string
+	dbcolomun     string
+	order_by      string
+	permission_id int64
+	action 		string
 }
 
-func (s *CrudController) auth(redirect string, w http.ResponseWriter, r *http.Request) {
+func (s *CrudController) hasPermission(dbtable, dbcolomun string, w http.ResponseWriter, r *http.Request) bool {
+	s.dbtable, s.dbcolomun = dbtable, dbcolomun
+	s.action = mux.Vars(r)["action"]
+
 	library.SESSION.Authentication(w, r)
-	if library.SESSION.GetSessionObj().GetAccount_id() == 0 {
-		if redirect != "" {
-			http.Redirect(w, r, redirect, http.StatusTemporaryRedirect)
-			return
-		}
+
+	if library.SESSION.GetSessionObj().GetAccountId() > 0 && library.SESSION.GetSessionObj().GetPositionId() > 0{
+		//if _, ok := library.POSITION.TreeMap[library.SESSION.GetSessionObj().GetPositionId()].GetPermissions()[s.permission_id]; ok {
+		//fmt.Println(library.SESSION.GetSessionObj())
+		library.VALIDATION.Status = 0
+		library.VALIDATION.Result = map[string]string{}
+
+		views.LOCALS.W = w
+		views.LOCALS.R = r
+		views.LOCALS.CurrentPage = s.action
+
+		return true
+		//}
 	}
-	library.VALIDATION.Status = 0
-	library.VALIDATION.Result = map[string]string{}
-}
-func (s *CrudController) authAdmin(dbtable string, w http.ResponseWriter, r *http.Request) string {
-	s.dbtable = dbtable
-	vars := mux.Vars(r)
-	action := vars["action"]
-	library.SESSION.Authentication(w, r)
-	if library.SESSION.GetSessionObj().GetPermission() != "admin" {
-		if action == "list" {
-			//w.Write([]byte(errors.ErrorView("Not permission","access denied")))
-			http.Redirect(w, r, "/", http.StatusTemporaryRedirect)
-			return ""
-		}
-		library.VALIDATION.Status = 80
-		library.VALIDATION.Result = map[string]string{"unauth":"access denied"}
-		return ""
-	}
-	library.VALIDATION.Status = 0
-	library.VALIDATION.Result = map[string]string{}
-	return action
+
+	library.VALIDATION.Status = 80
+	library.VALIDATION.Result = map[string]string{"unauth": "no permission"}
+	/*
+	if r.Method == "GET" {
+		http.Redirect(w, r, "/", http.StatusTemporaryRedirect)
+	}*/
+
+	return false
 }
 func (s *CrudController) getList(enable bool, w http.ResponseWriter, r *http.Request) (int64, string) {
 	return library.VALIDATION.IsInt64(false, "page", 20, r)
@@ -114,30 +116,50 @@ func (s *CrudController) getAjaxList(enable bool, w http.ResponseWriter, r *http
 }
 
 func (s *CrudController) get(w http.ResponseWriter, r *http.Request) (int64, string) {
-	idInt64, idStr := library.VALIDATION.IsInt64(true, s.dbtable + "_id", 20, r)
+	idInt64, idStr := library.VALIDATION.IsInt64(true, s.dbcolomun, 20, r)
 	return idInt64, idStr
 }
 func (s *CrudController) add(w http.ResponseWriter, r *http.Request) {
 
 }
 func (s *CrudController) edit(w http.ResponseWriter, r *http.Request) (int64, string) {
-	idInt64, idStr := library.VALIDATION.IsInt64(true, s.dbtable + "_id", 20, r)
+	idInt64, idStr := library.VALIDATION.IsInt64(true, s.dbcolomun, 20, r)
 	return idInt64, idStr
 }
 func (s *CrudController) del(w http.ResponseWriter, r *http.Request) (int64, string) {
-	idInt64, idStr := library.VALIDATION.IsInt64(true, s.dbtable + "_id", 20, r)
+	idInt64, idStr := library.VALIDATION.IsInt64(true, s.dbcolomun, 20, r)
 	return idInt64, idStr
 }
 
 func (s *CrudController) inlist(columns map[string]string, r *http.Request) {
+	if library.SESSION.GetSessionObj().GetAccountId() == 0 ||  library.SESSION.GetSessionObj().GetPositionId() == 0 {
+		return
+	}
+	var permission = ``
+	if s.dbtable == "position" || s.dbtable == "permission" {
+
+		if library.SESSION.GetSessionObj().GetPositionId() > 0 && library.POSITION.TreeMap[library.SESSION.GetSessionObj().GetPositionId()].GetParent().GetId() > 0 {
+			if library.POSITION.TreeMap[library.SESSION.GetSessionObj().GetPositionId()].GetDescendantIdsString() != `` {
+				if s.dbtable == "position" {
+					permission = ` AND position_id` + ` IN (` + library.POSITION.TreeMap[library.SESSION.GetSessionObj().GetPositionId()].GetDescendantIdsString() + `)`
+				}
+				if s.dbtable == "permission" {
+					permission = ` AND permission_position` + ` IN (` + library.POSITION.TreeMap[library.SESSION.GetSessionObj().GetPositionId()].GetDescendantIdsString() + `)`
+				}
+			} else {
+				return
+			}
+		}
+	}
 	r.ParseForm()
 	for columnName, columnType := range columns {
-		s.inlistUpdateQuery(columnName, columnType, r)
+		s.inlistUpdateQuery(columnName, columnType, permission, r)
+
+		s.inlistDeleteQuery(permission, r)
 	}
-	s.inlistDeleteQuery(r)
 }
 
-func (s *CrudController) inlistUpdateQuery(columnName, columnType string, r *http.Request) {
+func (s *CrudController) inlistUpdateQuery(columnName, columnType string, permission string, r *http.Request) {
 	var valueNum int = 1 // count numbers 1,2,3 ... to "$1,$2,$3,$4,$5 ..." string
 	var exec []interface{}
 	var layout_item_query = `UPDATE ` + s.dbtable + `
@@ -147,12 +169,14 @@ func (s *CrudController) inlistUpdateQuery(columnName, columnType string, r *htt
 			 unnest(ARRAY[ %s ]) AS v,
 			 unnest(ARRAY[ %s ]) AS i
 		     )AS data_table
-		WHERE ` + s.dbtable + `_id = data_table.i
+		WHERE ` + s.dbtable + `_id = data_table.i ` + permission + `
+
 		`
+
 	var values string
 	var ids string
 	var match bool
-	for _, v1 := range r.PostForm[columnName + "_inlist[]"] {
+	for _, v1 := range r.PostForm[columnName+"_inlist[]"] {
 		// The string - 25|89  or  25|253.56  or  25|Some Title
 		// is splitting to array [25,89] or [25,253.56]  or  [25,Some Title]
 		id_value_arr := strings.SplitN(v1, "|", 2)
@@ -188,20 +212,37 @@ func (s *CrudController) inlistUpdateQuery(columnName, columnType string, r *htt
 				ids += `,` + id_value_arr[0]
 				valueNum++
 			case "string":
+				if id_value_arr[1] != `` {
 					// building "$1,$2,$3,$4,$5 ..." string to values array: unnest(ARRAY[$1,$2,$3,$4,$5 ...]) AS v
 					values += `,$` + strconv.Itoa(valueNum)
 					exec = append(exec, template.HTMLEscapeString(id_value_arr[1]))
 					ids += `,` + id_value_arr[0]
 					valueNum++
-				
+				}
+			case "phone":
+				match, _ = regexp.MatchString(app.Pattern_phone(), id_value_arr[1])
+				if match {
+					values += `,$` + strconv.Itoa(valueNum)
+					exec = append(exec, id_value_arr[1])
+					ids += `,` + id_value_arr[0]
+					valueNum++
+				}
+			case "password":
+				if len(id_value_arr[1])>5 && len(id_value_arr[1])<128 {
+					values += `,$` + strconv.Itoa(valueNum)
+					exec = append(exec, library.SESSION.Cryptcode(id_value_arr[1]))
+					ids += `,` + id_value_arr[0]
+					valueNum++
+				}
 			}
 		}
 	}
 	if ids != `` {
+		//fmt.Printf(layout_item_query, values[1:], ids[1:])
 		model.Crud.Update(fmt.Sprintf(layout_item_query, values[1:], ids[1:]), exec)
 	}
 }
-func (s *CrudController) inlistDeleteQuery(r *http.Request) {
+func (s *CrudController) inlistDeleteQuery(permission string, r *http.Request) {
 	// delete in list
 	var del_ids string
 	for _, v2 := range r.PostForm["del_inlist[]"] {
@@ -217,12 +258,12 @@ func (s *CrudController) inlistDeleteQuery(r *http.Request) {
 	var del_q string
 	if del_ids != `` {
 		del_q = fmt.Sprintf(`
-		DELETE FROM ` + s.dbtable + ` WHERE ` + s.dbtable + `_id = ANY(ARRAY[%s])`, del_ids[1:])
+		DELETE FROM `+ s.dbtable+ ` WHERE `+ s.dbtable+ `_id = ANY(ARRAY[%s]) `+ permission, del_ids[1:])
 		if model.Crud.Delete(del_q, []interface{}{}) > 0 {
 			return
 		} else {
 			library.VALIDATION.Status = 200
-			library.VALIDATION.Result[s.dbtable + "_delete"] = "not deleted"
+			library.VALIDATION.Result[s.dbtable+"_delete"] = "not deleted"
 			return
 		}
 	}
